Interfaced all client and server side code.
Shifted what was previously the 'WhieboardModel' class so that it is now a
subclass of 'WhiteboardGUI'. Renamed the previous separate 'WhiteboardModel' class
to WhiteboardSeparateModel to avoid conflicts in WhiteboardGUI. Implemented
extra methods. Changed spec of WhiteboardGUI class so that it is testable on one machine.
Changed the main methods in both WhiteboardGUI and WhiteboardServer, also for testing purposes.